+++ .babelrc
@@ -5,5 +5,8 @@
     "babel-plugin-stylus-compiler",
     "@babel/plugin-proposal-function-bind"
   ],
-  "presets": ["@babel/preset-env", "@babel/preset-react"]
+  "presets": [
+    ["@babel/preset-env", { "useBuiltIns": "entry", "corejs": "3.26.1" }],
+    "@babel/preset-react"
+  ]
 }
